test: splits the JSON interpreter from run-spec-test.py into run-wast.py.
Adds --wast to run wast2json on the fly.
